// src/types.ts
export interface PlayerType {
    bats: string
    birth_date: string
    college: string
    end_date: string
    height_feet: string
    height_inches: string
    jersey_number: string
    name_display_first_last: string
    name_display_last_first: string
    name_first: string
    name_full: string
    name_last: string
    name_use: string
    player_id: string
    position_txt: string
    primary_position: string
    pro_debut_date: string
    start_date: string
    starter_sw: string
    status_code: string
    team_abbrev: string
    team_code: string
    team_id: string
    team_name: string
    throws: string
    weight: string
}

export interface HittingStatsType {
    ab: string
    ao: string
    avg: string
    babip: string
    bb: string
    cs: string
    d: string
    g: string
    gidp: string
    gidp_opp: string
    go: string
    go_ao: string
    h: string
    hbp: string
    hfly: string
    hgnd: string
    hldr: string
    hpop: string
    hr: string
    ibb: string
    lob: string
    np: string
    obp: string
    ops: string
    player_id: string
    ppa: string
    r: string
    rbi: string
    roe: string
    sac: string
    sb: string
    sf: string
    slg: string
    so: string
    sport: string
    sport_code: string
    sport_id: string
    t: string
    tb: string
    team_count: string
    tpa: string
    wo: string
    xbh: string
}

export interface PitchingStatsType {
    ab: string
    ao: string
    avg: string
    babip: string
    bb: string
    bb9: string
    bk: string
    bq: string
    bqs: string
    cg: string
    cs: string
    db: string
    er: string
    era: string
    g: string
    gf: string
    gidp: string
    gidp_opp: string
    go: string
    go_ao: string
    gs: string
    h: string
    h9: string
    hb: string
    hfly: string
    hgnd: string
    hld: string
    hldr: string
    hpop: string
    hr: string
    hr9: string
    ibb: string
    ip: string
    ir: string
    irs: string
    k9: string
    kbb: string
    l: string
    np: string
    obp: string
    ops: string
    pgs: string
    pip: string
    pk: string
    player_id: string
    ppa: string
    qs: string
    r: string
    rs9: string
    s: string
    sac: string
    sb: string
    sf: string
    sho: string
    slg: string
    so: string
    spct: string
    sport: string
    sport_code: string
    sport_id: string
    sv: string
    svo: string
    tbf: string
    team_count: string
    tr: string
    w: string
    whip: string
    wp: string
    wpct: string
}

export interface TeamTypes {
    team: string 
    team_id: string 
    gif: string
}